NN-3485 Rename viewAppointments
